+++ lib/hui/url.ex
@@ -185,6 +185,7 @@ defmodule Hui.URL do
   def encode_query(%Hui.H{} = url_params), do: encode_query(url_params |> Map.to_list)
   def encode_query(%Hui.H1{} = url_params), do: encode_query(url_params |> Map.to_list)
   def encode_query(%Hui.H2{} = url_params), do: encode_query(url_params |> Map.to_list)
+  def encode_query(%Hui.H3{} = url_params), do: encode_query(url_params |> Map.to_list |> Enum.sort)
 
   def encode_query(%Hui.F.Range{} = url_params), do: encode_query(url_params |> Map.to_list, "facet.range", url_params.range, url_params.per_field)
   def encode_query(%Hui.F.Interval{} = url_params), do: encode_query(url_params |> Map.to_list, "facet.interval", url_params.interval, url_params.per_field)
@@ -194,6 +195,7 @@ defmodule Hui.URL do
   def encode_query([{:__struct__, Hui.H} | tail]), do: Enum.map(tail, &prefix(&1, "hl")) |> encode_query
   def encode_query([{:__struct__, Hui.H1} | tail]), do: Enum.map(tail, &prefix(&1, "hl")) |> encode_query
   def encode_query([{:__struct__, Hui.H2} | tail]), do: Enum.map(tail, &prefix(&1, "hl")) |> encode_query
+  def encode_query([{:__struct__, Hui.H3} | tail]), do: Enum.map(tail, &prefix(&1, "hl")) |> encode_query
 
   def encode_query(enumerable) when is_list(enumerable), do: Enum.reject(enumerable, &invalid_param?/1) |> Enum.map_join("&", &encode/1)
   def encode_query(_), do: ""
@@ -232,6 +234,7 @@ defmodule Hui.URL do
       {:method, "facet.range"} -> {:"facet.range.method", v} # ditto
       {:interval, "facet.interval"} -> {:"facet.interval", v} # ditto
       {:per_field, _} -> {k, nil} # do not render this field
+      {:per_field_method, _} -> if per_field, do: {:"f.#{field}.#{prefix}.method", v}, else: {k, nil}
       {_, _} -> if per_field, do: {:"f.#{field}.#{prefix}.#{k}", v}, else: {:"#{prefix}.#{k}", v}
     end
   end